% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{align_var}
\alias{align_var}
\title{Align variables in time (R workflow)}
\usage{
align_var(
  x,
  variables = "temperature",
  offset = -0.5,
  interp_method = "linear",
  na_rm = FALSE
)
}
\arguments{
\item{x}{oce object}

\item{variables}{Character vector of data variable names to offset.}

\item{offset}{Numeric vector of offsets (in seconds) to add.}

\item{na_rm}{Remove scans with NAs in variable channel(s) after alignment.}

\item{method}{Interpolation method for the approx function to use.}
}
\value{
Returns an oce object with offsets applied.
}
\description{
Offset variables in time to account for lags in sensor responses. Variables are interpolated so offsets do not need to be multiples of the scan interval.
}
