% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gapctd.R
\name{append_haul_data}
\alias{append_haul_data}
\title{Find cast times and metadata for a haul (R workflow)}
\usage{
append_haul_data(x, haul_df, ctd_tz = "America/Anchorage")
}
\arguments{
\item{x}{oce object}

\item{haul_df}{data.frame containing haul metadata}

\item{ctd_tz}{timezone for the ctd as a character vector or numeric}
}
\value{
A data.frame with haul metadata and cast times.
}
\description{
Find haul metadata for a file and appends to oce object.
}
