% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{assign_metadata_fields}
\alias{assign_metadata_fields}
\title{Assign metadata fields (R workflow)}
\usage{
assign_metadata_fields(x, cast_direction)
}
\arguments{
\item{x}{oce object that contains}

\item{cast_direction}{Cast direction ("downcast", "upcast", "bottom)}
}
\value{
oce object with metadata fields updated with latitude, longitude, ship, bottom depth, deployment type, cruise, date, institute, scientist.
}
\description{
Assign metadata fields from x@metadata@race_metadata to the object.
}
