% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{bin_average}
\alias{bin_average}
\title{Bin average: Average variables by depth/pressure bin}
\usage{
bin_average(
  x,
  by = "depth",
  bin_width = 1,
  exclude_surface = 0.5,
  exclude_bad_flag = TRUE,
  interpolate_missing = TRUE,
  missing_latitude = 55
)
}
\arguments{
\item{x}{ctd object}

\item{by}{Grouping variable as a character vector ("depth" or "pressure")}

\item{bin_width}{Width of bins to use for binning.}

\item{exclude_surface}{Depth or pressure (down is positive) as numeric vector (1L). Scans will be excluded from binning if they are shallower this threshold.}

\item{exclude_bad_flag}{Logical indicating whether to remove scans with a bad (negative) flag.}

\item{interpolate_missing}{Interpolate variables when there are no good data for a depth/pressure bin.}

\item{missing_latitude}{Latitude in decimal degrees to use for depth estimation if ctd object does not contain latitude.}
}
\value{
ctd object with variables binned and 'flag' column removed.
}
\description{
Calculate averages for variables by depth or pressure bin.
}
\author{
Sean Rohan
}
