% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bin_summary.R
\name{calc_bin_summary}
\alias{calc_bin_summary}
\title{Calculate variable averages from accepted profiles, by depth/pressure bin (SBEDP workflow)}
\usage{
calc_bin_summary(
  accepted_dir = here::here("output", "accepted_profiles"),
  z_var = "pressure",
  z_range = 1,
  agg_fun = mean
)
}
\arguments{
\item{accepted_dir}{File path to folder with data from accepted profiles.}

\item{z_var}{Name of the z variable ("pressure" or "depth").}

\item{agg_fun}{Function to use.}

\item{depth_range}{Range of depth/pressure bins to use, as a 1L or 2L numeric vector.}
}
\description{
Calculate temperature from the 1 db pressure bin.
}
