% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_bottom_mean}
\alias{calc_bottom_mean}
\title{Calculate station bottom temperature and salinity}
\usage{
calc_bottom_mean(
  haul_metadata_path = list.files(paste0(getwd(), "/metadata/"), full.names = TRUE),
  pattern = "TEOS10.cnv",
  timezone = "America/Anchorage",
  time_buffer = 30,
  append_haul_metadata = TRUE
)
}
\arguments{
\item{haul_metadata_path}{Character vector. Path to haul data .csv generated by get_bottom_data().}

\item{pattern}{Character vector. Suffix of cnv file to search for. Default = "TEOS10.cnv"}

\item{timezone}{Timezone as a character vector to force instrument time. Default = "America/Anchorage" to correspond with survey time.}

\item{time_buffer}{1L numeric vector. Buffer between on-bottom/off-bottom events and first/last data point to use to calculate bottom values, in seconds.}

\item{append_haul_metadata}{Logical. Should the file named in haul_metadata_path be appended (overwritten) to include mean bottom values and metadata?}
}
\description{
Use the OCE package to retrieve CTD bottom data from cnv files and calculate mean temperature and salinity.
}
