% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_netcdf.R
\name{calc_bottom_mean_oce}
\alias{calc_bottom_mean_oce}
\title{Calculate means for an oce object}
\usage{
calc_bottom_mean_oce(x, prefix = "mean_bottom_", fun = mean, vars = NULL)
}
\arguments{
\item{x}{oce object}

\item{prefix}{Character vector to append to variable names.}

\item{fun}{Function to apply to variable vectors.}

\item{vars}{Variables to apply function to. Automatically assigned when NULL (default)}
}
\value{
A data.frame with variables calculated and race_metadata if provided (x@metadata$race_metadata)
}
\description{
Apply function to selected variables in an oce object and return a data frame.
}
\author{
Sean Rohan
}
