% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fixed_depth_ts.R
\name{calc_fixed_depth_ts}
\alias{calc_fixed_depth_ts}
\title{Calculate fixed depth temperature and salinity (SBEDP workflow)}
\usage{
calc_fixed_depth_ts(
  haul_metadata_path = list.files(paste0(getwd(), "/metadata/"), full.names = TRUE),
  depth_interval = 1,
  cast_direction = "upcast",
  append_haul_metadata = TRUE
)
}
\arguments{
\item{haul_metadata_path}{Character vector. Path to haul data .csv generated by get_bottom_data().}

\item{cast_direction}{Character vector, either "upcast" or "downcast."}

\item{append_haul_metadata}{Logical. Should the file named in haul_metadata_path be appended (overwritten) to include mean surface values and metadata?}

\item{depth}{Numeric vector (1L or 2L) indicating a fixed depth or range of depths from which to calculate a mean.}
}
\description{
Use the OCE package to retrieve CTD near-surface data from upcast or downcast cnv files and calculate mean temperature and salinity for a fixed depth. Run after calc_bottom_mean().
}
