% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_surface_mean.R
\name{calc_surface_mean}
\alias{calc_surface_mean}
\title{Calculate station surface temperature and salinity}
\usage{
calc_surface_mean(
  haul_metadata_path = list.files(paste0(getwd(), "/metadata/"), full.names = TRUE),
  depth_interval = c(2, 5),
  append_haul_metadata = TRUE
)
}
\arguments{
\item{haul_metadata_path}{Character vector. Path to haul data .csv generated by get_bottom_data().}

\item{append_haul_metadata}{Logical. Should the file named in haul_metadata_path be appended (overwritten) to include mean surface values and metadata?}

\item{depth_inteval}{Range of depths to use to calculate means.}
}
\description{
Use the OCE package to retrieve CTD near-surface data from downcast cnv files and calculate mean surface temperature and salinity. Run after calc_bottom_mean().
}
