% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_psa_parameters.R
\name{change_psa_alignctd}
\alias{change_psa_alignctd}
\title{Change AlignCTD parameters (SBEDP workflow)}
\usage{
change_psa_alignctd(variable, alignment, psa_path = NULL)
}
\arguments{
\item{alignment}{Adjustment to channel alignment relative to pressure, in seconds. The 'typical' alignment for a pumped SBE19plus is 0.5 for temperature.}

\item{psa_path}{File path to the alignctd psa file. If not provided, attempts to use /psa_xmlcon/alignctd.psa.}

\item{var}{Character vector indicating which variable channel should be aligned ("temperature", "conductivity", or "oxygen")}
}
\description{
Change SBE Data Processing alignment module parameters in a .psa file.
}
\details{
This funtion modifies text in psa files that are used to run the AlignCTD module in SBE Data Processing.
}
