% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_psa_parameters.R
\name{change_psa_celltm}
\alias{change_psa_celltm}
\title{Change CellTM parameters}
\usage{
change_psa_celltm(
  alpha,
  tau_ctm,
  psa_path = here::here("psa_xmlcon", "celltm.psa")
)
}
\arguments{
\item{alpha}{Alpha parameter. 1L numeric vector.}

\item{tau_ctm}{Tau parameter (i.e., inv_beta) 1L numeric vector.}

\item{psa_path}{File path to cell thermal mass .psa file.}
}
\description{
Change SBE Data Processing conductivity cell thermal inertia correction module parameters in a .psa file.
}
\details{
This funtion modifies text in psa files that are used to run the CellTM module in SBE Data Processing.
}
