% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{channel_correlation}
\alias{channel_correlation}
\title{Correlation between two channels (R workflow)}
\usage{
channel_correlation(
  x,
  exclude_flag = TRUE,
  c1 = "temperature",
  c2 = "conductivity",
  min_pressure = 4,
  cor_method = "pearson"
)
}
\arguments{
\item{x}{oce object}

\item{exclude_flag}{Logical denoting whether bad scans (flag < 0) should be excluded.}

\item{c1}{Name of first channel as a character vector. Default = "temperature"}

\item{c2}{Name of second channel as a character vector. Default = "conductivity"}

\item{min_pressure}{Minimum pressure to use for correlations}
}
\value{
Correlation coefficient (1L numeric)
}
\description{
Calculate correlation coefficient between two channels (e.g. temperature and conductivity)
}
