% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_density_inversion.R
\name{check_density_inversion}
\alias{check_density_inversion}
\title{Check density inversion (R workflow)}
\usage{
check_density_inversion(
  x,
  threshold = -1e-05,
  threshold_method = "bv",
  correct_inversion = TRUE
)
}
\arguments{
\item{x}{oce object}

\item{threshold}{Numerical. Threshold for flagging a density inversion. Must be negative. Default (-1e-5) is the buoyancy frequency threshold PMEL uses. GTSPP uses a density threshold of 0.05.}

\item{threshold_method}{Character. "density" to flag inversions based on density anomalies, "bv" for Brunt-Vaisala buoyancy frequency}

\item{correct_inversion}{Logical. Should density inversions be removed and interpolated?}
}
\value{
oce object with density inversions flaggd and/or interpolated
}
\description{
Check density inversion (R workflow)
}
\author{
Sean Rohan
}
