% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{compare_methods}
\alias{compare_methods}
\title{Compare methods based on spikes and make multipanel plots (SBEDP workflow)}
\usage{
compare_methods(
  prefix,
  processing_method,
  method_labels = NULL,
  return_output = FALSE,
  pattern_downcast = "downcast.cnv",
  pattern_upcast = "upcast.cnv",
  min_pressure_bin = 4,
  area_method = "ts",
  scale_vars = FALSE
)
}
\arguments{
\item{prefix}{Character vector to include before output plots.}

\item{processing_method}{Character vector denoting the name of the processing method, corresponding with names of folders in /output/ that contain processed cnv files.}

\item{method_labels}{Character vector denoting titles to use on multipanel plots. Must be the same length as processing_method. Default (NULL) uses processing_method character vector for plot titles.}

\item{return_output}{Logical. If true, returns output in a data.frame.}

\item{pattern_downcast}{Character vector pattern for downcast file.}

\item{pattern_upcast}{Character vector pattern for upcast file.}

\item{min_pressure_bin}{Numeric vector (1L) indicating the minimum depth bin to use for comparing profiles.}

\item{area_method}{Should area calculations use temperature-salinity ("ts") or pressure-salinity ("ps") profiles?}

\item{scale_vars}{Logical. Should temperature and salinity be scaled to calculate T-S area?}
}
\description{
Compare processing methods based on spikiness of profiles and make multipanel plots to compare methods/workflows for processing CTD data.
}
