% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{convert_ctd_hex}
\alias{convert_ctd_hex}
\title{Convert hex files using SBE Data Processing software}
\usage{
convert_ctd_hex(
  hex_file_path,
  xmlcon_path,
  bat_file = NULL,
  datcnv_file = NULL
)
}
\arguments{
\item{xmlcon_path}{Path to xmlcon file}

\item{ctd_directory_path}{Path to CTD directory}
}
\description{
Converts .hex files to .cnv files using SBE Data Processing. Must have SBE Data Processing software installed for this function to work.
}
\examples{
# Copy system files to working directory for example
file.copy(from = system.file("exdata/CTDtoBTD/2021_06_13_0003.hex", package = "GAPsurvey"),
to = gsub(pattern = system.file("exdata/CTDtoBTD/", package = "GAPsurvey"),
          replacement = getwd(),
          x = system.file("exdata/CTDtoBTD/2021_06_13_0003.hex", package = "GAPsurvey")))

file.copy(from = system.file("exdata/CTDtoBTD/19-8102_Deploy2021.xmlcon", package = "GAPsurvey"),
          to = gsub(pattern = system.file("exdata/CTDtoBTD/", package = "GAPsurvey"),
                    replacement = getwd(),
                    x = system.file("exdata/CTDtoBTD/19-8102_Deploy2021.xmlcon", package = "GAPsurvey")))

# Run convert_ctd_hex()
GAPsurvey:::convert_ctd_hex(hex_file_path = "2021_06_13_0003.hex",
                            xmlcon_path = "19-8102_Deploy2021.xmlcon",
                            bat_file = NULL,
                            datcnv_file = NULL)
}
