% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{convert_do_to_o2sat}
\alias{convert_do_to_o2sat}
\title{Calculate oxygen saturation (percent) from dissolved oxygen (ml/l)}
\usage{
convert_do_to_o2sat(oxygen, temperature, salinity)
}
\arguments{
\item{oxygen}{Dissolved oxygen in ml/l}

\item{temperature}{Temperature (IPTS-68, degrees Celsius).}

\item{salinity}{Salinity (PSU, PSS-78).}
}
\description{
Dissolved oxygen divided by oxygen saturation calculated following Garcia and Gordon (1992)
}
\references{
Garcia, H.E., Gordon, L.I., 1992. Oxygen solubility in seawater: Better fitting equations. Limnol. Oceanogr. 37, 1307–1312. https://doi.org/10.4319/lo.1992.37.6.1307
}
