% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_density_inversion.R
\name{correct_density_inversion}
\alias{correct_density_inversion}
\title{Flag density inversion and correct (SBEDP workflow)}
\usage{
correct_density_inversion(
  threshold = NULL,
  threshold_method = "bv",
  correct_inversion = TRUE,
  cnv_dir = NULL,
  ...
)
}
\arguments{
\item{threshold}{Numerical. Threshold for flagging a density inversion. Must be negative. Default (-1e-5) is the buoyancy frequency threshold PMEL uses. GTSPP uses a density threshold of 0.05.}

\item{threshold_method}{Character. "density" to flag inversions based on density anomalies, "bv" for Brunt-Vaisala buoyancy frequency}

\item{correct_inversion}{Logical. Should density inversions be removed and interpolated?}

\item{cnv_dir}{File path to cnv directory.}
}
\value{
Writes flagged and corrected data to csvs in output/density_corrected.
}
\description{
Screen profiles for density inversions and flag based on a threshold criteria based on density (kg m^-3) or Brunt-Vaisala buoyancy frequency (N^2), s^-2. Either  Optional: remove missing value and interpolate.
}
