% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NMEA_files.R
\name{create_NMEA_files}
\alias{create_NMEA_files}
\title{Function to pull in lat/lon from bottom trawl racebase haul data that matches each .cnv to create NMEA files}
\usage{
create_NMEA_files()
}
\arguments{
\item{oracle}{password Enter your personal oracle password to access RACEBASE.HAUL oracle data}

\item{survey}{year Specify the survey year of your CTD data. Prompted, no default.}
}
\description{
implement between derive.bat and getdata.bat
requires VPN connection
will prompt for oracle username, password, survey region (GOA, AI, BS), and survey year
dependencies: getPass, RODBC, oce
}
