% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm_functions.R
\name{ctm_adjust}
\alias{ctm_adjust}
\title{Perform conductivity cell thermal intertia correction and compare upcasts to downcasts}
\usage{
ctm_adjust(
  alpha = 0.04,
  beta = 0.125,
  f_n = 0.25,
  down_temperature,
  down_pressure,
  down_conductivity,
  up_temperature,
  up_pressure,
  up_conductivity,
  min_pressure_bin = 0,
  buffer_bottom_pressure_bin = 0,
  obj_fn = NULL
)
}
\arguments{
\item{f_n}{Numeric vector (1L). Scan interval in seconds. Default = 0.25 for SBE19plus default 4 Hz scan interval.}

\item{down_temperature}{Numeric vector of downcast temperatures in degC (ITS-90 scale),}

\item{down_pressure}{Numeric vector of downcast pressure in dbar.}

\item{down_conductivity}{Numeric vector of downcast conductivity in S/m.}

\item{up_temperature}{Numeric vector of upcast temperatures in degC (ITS-90 scale),}

\item{up_pressure}{Numeric vector of upcast pressure in dbar.}

\item{up_conductivity}{Numeric vector of upcast conductivity in S/m.}

\item{min_pressure_bin}{Numeric vector (1L) indicating the minimum depth bin to use for calculating the objective function. Default = 2 dbar to minimize surface artifacts.}

\item{buffer_bottom_pressure_bin}{Numeric vector (1L). Buffer for excluding near-bottom pressure bins to minimize effects of depth changes on objective function.}

\item{obj_fn}{Logical. Return the objective function instead of corrected conductivity, temperature, etc.}

\item{alpha_beta}{Numeric vector (2L). Log-transformed alpha and beta parameters in thermal mass correction algorithm (alpha default = 0.04 and beta default = 0.125 for SBE19plus), passed to ctm_par_a() and ctm_par_b()}
}
\description{
Calculates conductivity cell thermal intertia correction using ctm_par_a(), ctm_par_b(), and ctm_correct_c_t(), and recalculates salinity for paired upcasts and downcasts. If argument obj_fn is set to difference or area, the function is used to calculate and return the value of an objective function which is mean absolute difference in salinity (PSS-78 scale) between upcast and downcast from the means of 1 dbar bins. If min_obj = FALSE, the function returns a data frame containing corrected values averaged by 1 dbar bins.
}
