% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{ctm_adjust_path}
\alias{ctm_adjust_path}
\title{Calculate conductivity cell thermal intertia correction (SBEDP workflow)}
\usage{
ctm_adjust_path(
  alpha = 0.04,
  tau = 8,
  f_n = 0.25,
  temperature,
  pressure,
  conductivity,
  flag,
  obj_fn = NULL,
  binned = TRUE
)
}
\arguments{
\item{alpha}{Numeric vector (1L). Alpha parameter in the cell thermal mass correction filter (default = 0.04 for SBE19plus)}

\item{tau}{Numeric vector (1L). Tau (inverse beta) parameter in the cell thermal mass correction filter (default = 8 for SBE19plus).}

\item{f_n}{Numeric vector (1L). Scan interval in seconds. Default = 0.25 for SBE19plus default 4 Hz scan interval.}

\item{temperature}{Numeric vector of temperature in degC (ITS-90 scale),}

\item{pressure}{Numeric vector of pressure in dbar.}

\item{conductivity}{Numeric vector of conductivity in S/m.}

\item{flag}{Character or numeric vector of data quality flags where NAs are scans to be excluded from final profiles. NA scans are used to estimate thermal cell inertia conductivity corrections and salinity but re removed from binned outputs.}

\item{obj_fn}{Logical. If "path", returns an objective function, the total path distance of the salinity profile. Otherwise, returns temperature, salinity, and conductivity in 1 db depth bins.}

\item{binned}{Logical. If TRUE, objective function is calculated for pressure-binned data. If FALSE, the objective function is calculated using all un-flagged data.}
}
\description{
Calculates conductivity cell thermal intertia correction based on alpha, inverse beta (tau[CTM]), scan interval, temperature, and conductivity, then recalculates salinity.
}
