% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{ctm_adjust_tsarea}
\alias{ctm_adjust_tsarea}
\title{Perform conductivity cell thermal inertia correction and compare upcasts to downcasts (SBEDP workflow)}
\usage{
ctm_adjust_tsarea(
  alpha = 0.04,
  tau = 8,
  f_n = 0.25,
  down_temperature,
  down_pressure,
  down_conductivity,
  down_flag,
  up_temperature,
  up_pressure,
  up_conductivity,
  up_flag,
  obj_fn = "none",
  scale_vars = FALSE
)
}
\arguments{
\item{alpha}{Numeric vector (1L). Alpha parameter in the cell thermal mass correction filter (default = 0.04 for SBE19plus)}

\item{tau}{Numeric vector (1L). Tau (inverse beta) parameter in the cell thermal mass correction filter (default = 8 for SBE19plus).}

\item{f_n}{Numeric vector (1L). Scan interval in seconds. Default = 0.25 for SBE19plus default 4 Hz scan interval.}

\item{down_temperature}{Numeric vector of downcast temperatures in degC (ITS-90 scale),}

\item{down_pressure}{Numeric vector of downcast pressure in dbar.}

\item{down_conductivity}{Numeric vector of downcast conductivity in S/m.}

\item{up_temperature}{Numeric vector of upcast temperatures in degC (ITS-90 scale),}

\item{up_pressure}{Numeric vector of upcast pressure in dbar.}

\item{up_conductivity}{Numeric vector of upcast conductivity in S/m.}

\item{obj_fn}{Character vector ("none" or "area")}

\item{scale_vars}{Logical. Should temperature and salinity be scaled to calculate T-S area?}
}
\description{
Calculates conductivity cell thermal intertia correction using ctm_par_a(), ctm_par_b(), and ctm_correct_c_t(), and recalculates salinity for paired upcasts and downcasts. If argument obj_fn is set to area, the function calculates and return the value of an objective function which is the area between upcast and downcast temperature-salinity curves in 1 dbar pressure bins. If obj_fn = "none," returns a data.frame with upcast and downcast temperature, conductivity, corrected conductivity, and salinity in 1 dbar pressure bins.
}
