% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm_functions.R
\name{ctm_correct_c_t}
\alias{ctm_correct_c_t}
\title{Conductivity thermal intertia correction factor (C[T])}
\usage{
ctm_correct_c_t(a, b, temperature)
}
\arguments{
\item{a}{Numeric vector (1L). a parameter in thermal mass correction formula, as calculated by gapctd::ctm_par_a()}

\item{b}{Numeric vector (1L). b parameter in thermal mass correction formula, as calculated by gapctd::ctm_par_b().}

\item{temperature}{Numeric vector of temperatures in degC (ITS-90 scale).}
}
\description{
Calculate thermal inertia adjustment factors for SBE19 conductivity cell using temperature and parameters a and b, using the same method as the Thermal Mass Correction module in SBE Data Processing software.
}
