% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm_functions.R
\name{ctm_par_b}
\alias{ctm_par_b}
\title{Conductivity thermal inertia correction parameter b}
\usage{
ctm_par_b(alpha, a)
}
\arguments{
\item{alpha}{Numeric vector (1L). Alpha parameter in thermal mass correction function (default = 0.04 for SBE19plus).}

\item{a}{Numeric vector (1L). Conductivity thermal mass correction value a.}
}
\description{
Calculate Sea-Bird conductivity cell thermal mass correction parameter b, which is used to apply thermal inertia correction to conductivity measurements.
}
