% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{derive_eos}
\alias{derive_eos}
\title{Derive EOS and GSW}
\usage{
derive_eos(x, precision = NULL)
}
\arguments{
\item{x}{oce object}

\item{precision}{Optional named numeric vector with precision (decimal places) for variables.}
}
\description{
Derive salinity, absolute salinity, sound speed, density, and buoyancy frequency.
}
