% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{derive_oxygen}
\alias{derive_oxygen}
\title{Apply tau correction to oxygen voltage channel}
\usage{
derive_oxygen(
  x,
  cal_rds_path = NULL,
  tau_correction = TRUE,
  aa = NA,
  bb = NA,
  cc = NA,
  ee = NA,
  soc = NA,
  Voffset = NA,
  tau20 = NA,
  d0 = NA,
  d1 = NA,
  d2 = NA,
  sig_digits = 4
)
}
\arguments{
\item{x}{oce object containing raw oxygen voltage data (rawOxygen).}

\item{cal_rds_path}{Optional path to calibration parameter file.}

\item{tau_correction}{Should the tau correction (Edwards et al., 2010) be used to account for time-dynamic errors in oxygen?}

\item{aa}{Calibration parameter A}

\item{bb}{Calibration parameter B}

\item{cc}{Calibration parameter C}

\item{ee}{Calibration parameter E}

\item{soc}{Calibration parameter Soc}

\item{Voffset}{Calibration parameter voltage offset}

\item{tau20}{Tau correction calibration parameter Tau20.}

\item{d0}{Tau correction calibration parameter D0.}

\item{d1}{Tau correction calibration parameter D1.}

\item{d2}{Tau correction calibration parameter D2.}

\item{sig_digits}{Significant digits for oxygen (ml/l).}
}
\value{
oce object with tau correction applied to oxygen voltage channel (oxygenRaw). Returns the input unchanged if no oxygen channel is detected.
}
\description{
Tau correction following Edwards et al. (2010). Should be run after dynamic corrections to temperature and conductivity channels.
}
\references{
Edwards, B., Murphy, D., Janzen, C., Larson, A.N., 2010. Calibration, response, and hysteresis in deep-sea dissolved oxygen measurements. J. Atmos. Ocean. Technol. 27, 920–931. https://doi.org/10.1175/2009JTECHO693.1
Garcia, H.E., Gordon, L.I., 1992. Oxygen solubility in seawater: Better fitting equations. Limnol. Oceanogr. 37, 1307–1312. https://doi.org/10.4319/lo.1992.37.6.1307
}
\author{
Sean Rohan
}
