% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_calibration_xmlcon}
\alias{extract_calibration_xmlcon}
\title{Extract calibration parameters from xmlcon to a list}
\usage{
extract_calibration_xmlcon(xmlcon_path)
}
\arguments{
\item{xmlcon_path}{Path to an xmlcon file.}
}
\description{
Retrives calibration parameters for temperature, conductivity, pressure, oxygen, and pH sensors from an instrument configuration (.xmlcon) file.
}
