% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finalize_data}
\alias{finalize_data}
\title{Move good and bad profiles to final directories (R workflow)}
\usage{
finalize_data(rds_dir_path = here::here("output", processing_method))
}
\arguments{
\item{rds_dir_path}{Path to directory containing rds files}
}
\description{
Move good and bad profiles to final directories (R workflow)
}
