% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{fixed_alignment}
\alias{fixed_alignment}
\title{Calculate alignment correlation for fixed parameters}
\usage{
fixed_alignment(
  x,
  align_pars,
  cor_var = "conductivity",
  cor_method = "pearson",
  cast_direction
)
}
\arguments{
\item{x}{oce object}

\item{align_pars}{A list object with alignment parameters for a variable, e.g., list(temperature = -0.5)}

\item{cor_var}{A vector of variable names that channels being aligned should be compared with (e.g., corr_var = "conductivity")}

\item{cor_method}{Correlation method, passed to gapctd::channel_correlation().}

\item{cast_direction}{Cast direction as a character vector ("downcast" or "upcast"). Passed to gapctd::slowdown()}
}
\description{
Calculate alignment correlation for fixed parameters
}
\author{
Sean Rohan
}
