% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{get_calibration_parameter}
\alias{get_calibration_parameter}
\title{Extract variable from xml}
\usage{
get_calibration_parameter(
  header,
  cal_par,
  start_block = NULL,
  end_block = NULL,
  make_numeric = TRUE
)
}
\description{
Extract variable from xml
}
