% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gapctd.R
\name{get_haul_data}
\alias{get_haul_data}
\title{Get haul data from RACEBASE (R workflow)}
\usage{
get_haul_data(
  channel,
  vessel,
  cruise,
  out_path = NULL,
  tzone = "America/Anchorage"
)
}
\arguments{
\item{channel}{RODBC connection.}

\item{vessel}{Vessel code as a 1L numeric vector.}

\item{cruise}{Cruise code as a numeric vector (can include more than one).}

\item{out_path}{Optional. Filepath for the output R data file (.rds).}

\item{tzone}{Time zone for events and star_time.}
}
\value{
Returns a data.frame containing haul data. Also saves haul data to an rds file.
}
\description{
Retrieves vessel, cruise, haul, event times, locations, bottom depth, gear temperature, surface temperature, performance, haul type, and haul start time from racebase and race_data.
}
