% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{get_haul_events}
\alias{get_haul_events}
\title{Retrieve haul event times from RACEBASE (SBEDP workflow)}
\usage{
get_haul_events(
  haul_metadata_path = list.files(paste0(getwd(), "/metadata/"), full.names = TRUE),
  append_haul_metadata = TRUE,
  channel
)
}
\arguments{
\item{haul_metadata_path}{Character vector. Path to haul data .csv generated by create_NMEA().}

\item{append_haul_metadata}{Logical. Should the file named in haul_metadata_path be appended (overwritten) to include haul events?}

\item{channel}{An RODBC connection to Oracle.}
}
\description{
Retrieves on bottom, haul back, off bottom, and marked event times from RACE_DATA.
}
