% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{integer_to_conductivity}
\alias{integer_to_conductivity}
\title{Convert SBE integer to conductivity}
\usage{
integer_to_conductivity(
  conductivity_integer,
  temperature,
  pressure,
  condg,
  condh,
  condi,
  condj,
  cpcor,
  ctcor,
  par0 = 256,
  par1 = 1000,
  sig_figs = 6
)
}
\description{
Convert SBE integer to conductivity
}
