% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{integer_to_pressure}
\alias{integer_to_pressure}
\title{Convert SBE integer to pressure}
\usage{
integer_to_pressure(
  temperature_integer,
  pressure_integer,
  tvoltage_integer,
  ptempa0,
  ptempa1,
  ptempa2,
  ptca0,
  ptca1,
  ptca2,
  ptcb0,
  ptcb1,
  ptcb2,
  pa0,
  pa1,
  pa2,
  par0 = 13107,
  sig_figs = 3,
  convert_to_dbar = TRUE
)
}
\description{
Convert SBE integer to pressure
}
