% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{integer_to_temperature}
\alias{integer_to_temperature}
\title{Convert SBE integer to ITS-90 temperature}
\usage{
integer_to_temperature(
  temperature_integer,
  sig_figs = NULL,
  a0,
  a1,
  a2,
  a3,
  offset = 0,
  par0 = 524288,
  par1 = 1.6e+07,
  par2 = 2.9e+09,
  par3 = 102400000,
  par4 = 20480,
  par5 = 2e+05
)
}
\description{
Convert SBE integer to ITS-90 temperature
}
