% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{loop_edit}
\alias{loop_edit}
\title{Flag slowdowns and reversals (R workflow)}
\usage{
loop_edit(x, min_speed = 0.1, window = 5, cast_direction = NULL)
}
\arguments{
\item{x}{oce object}

\item{min_speed}{Numeric vector (1L). Threshold below which observations should be flagged.}

\item{window}{Numeric vector (1L). Number of samples to use to calculate speeds for each observation.}

\item{cast_direction}{Character vector ("downcast", "upcast") denoting the cast direction.}
}
\value{
oce object with flags updated to denote scans for which speeds were below the threshold.
}
\description{
Flags slowdowns and reversals in profiles based on minimum speed.
}
