% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{lowpass_filter}
\alias{lowpass_filter}
\title{Low-pass filter (R workflow)}
\usage{
lowpass_filter(
  x,
  variables = c("temperature", "conductivity", "pressure"),
  time_constant = c(0.5, 0.5, 1),
  precision = c(4, 6, 3),
  freq_n = 0.25
)
}
\arguments{
\item{x}{oce object}

\item{variables}{Character vector of data variable names to filter.}

\item{time_constant}{Numeric vector of time constants for filters (in seconds).}

\item{freq_n}{Optional. Sampling interval in seconds (1L numeric). Sampling interval is inferred from timeS if not provided.}
}
\description{
Low-pass filter variables based on time using the filter from SBE data processing.
}
