% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cast_plots.R
\name{make_cast_plots}
\alias{make_cast_plots}
\title{Make profile plots from .cnv files}
\usage{
make_cast_plots(plot_bat = NA)
}
\arguments{
\item{plot_bat}{Name of the batch file to use. Uses the default file if not provided.}
}
\description{
Plots for visual data QA/QC from .cnv files. Reads data from final_cnv file.
}
