% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cast_sections.R
\name{make_cast_sections}
\alias{make_cast_sections}
\title{Make sections from upcasts and downcasts}
\usage{
make_cast_sections(
  haul_metadata_path = list.files(paste0(getwd(), "/metadata/"), full.names = TRUE),
  pattern = "TEOS10.cnv",
  xmlcon_file = NA,
  section_bat = NA,
  pressure_bat = NA,
  binavg_bat = NA
)
}
\arguments{
\item{haul_metadata_path}{Character vector. Path to haul data .csv generated by get_bottom_data().}

\item{pattern}{Character vector. Suffix of cnv file to search for. Default = "TEOS10.cnv"}

\item{xmlcon_file}{xmlcon file}

\item{section_bat}{section batch file}

\item{pressure_bat}{pressure batch file}

\item{binavg_bat}{binavg batch file}
}
\description{
Split cnv file data into upcast and downcast files based on haul event times using the SBEDataProcessing 'section' module. This is used because SBEDataProcessing split methods result in profiles that include samples while the trawl was being towed along the bottom, which results in sampling artifacts in the bottom depth bin.
}
