% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_netcdf.R
\name{make_ctd_ncdf}
\alias{make_ctd_ncdf}
\title{Make netCDF from profiles and metadata}
\usage{
make_ctd_ncdf(
  fpath = c(list.files(path = here::here("output", "accepted_profiles"), full.names =
    TRUE), list.files(path = "C:/Users/sean.rohan/Work/afsc/WIP/AKK/accepted_profiles",
    full.names = TRUE)),
  metadata_path = c(list.files(path = here::here("metadata"), full.names = TRUE),
    list.files(path = "C:/Users/sean.rohan/Work/afsc/WIP/AKK/metadata", full.names =
    TRUE)),
  output_file = "output.nc",
  global_attributes = list(citation = "CTD TEAM...", id = "doi", cdm_data_type =
    "Point", cruise =
    "2021 Eastern Bering Sea Continental Shelf and Northern Bering Sea Bottom-Trawl Survey",
    institution = "NOAA Alaska Fisheries Science Center", contributor_name = "",
    creator_name = "", creator_institution = "NOAA Alaska Fisheries Science Center",
    creator_email = "", publisher = "NOAA Alaska Fisheries Science Center",
    publisher_type = "institution", publisher_url =
    "https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center",     
    geospatial_bounds_crs = "EPSG:4326", license =
    "These data may be redistributed and used without restriction.", metadata_link =
    "[DOI]", instrument = "CTD", standard_name_vocabulary = "CF Standard Name Table v79",
    source = paste0("CTD data processed using gapctd ", packageVersion(pkg = "gapctd")))
)
}
\arguments{
\item{fpath}{Path to accepted profiles.}

\item{metadata_path}{Path to metadata files.}

\item{global_attributes}{List of global attributes that is passed to gapctd::df_to_netcdf(global_attributes)}
}
\description{
Read-in files from /accepted_profiles/ and /metadata/ subdirectories and create a netCDF file. Combines data from multiple vessels from the same cruise.
}
