% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{make_ctd_zip}
\alias{make_ctd_zip}
\title{Package final files and metadata in a zip file (SBEDP workflow)}
\usage{
make_ctd_zip(region, vessel, year)
}
\arguments{
\item{region}{Survey region}

\item{vessel}{Survey vessel}

\item{year}{Survey year}
}
\description{
Package final files and metadata in a zip file (SBEDP workflow)
}
