% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_method_multipanel_plot.R
\name{make_method_multipanel_plot}
\alias{make_method_multipanel_plot}
\title{Multipanel plots to compare methods}
\usage{
make_method_multipanel_plot(
  prefix,
  processing_method,
  method_labels = NULL,
  return_output = FALSE
)
}
\arguments{
\item{prefix}{Character vector to include before output plots.}

\item{processing_method}{Character vector denoting the name of the processing method, corresponding with names of folders in /output/ that contain processed cnv files.}

\item{method_labels}{Character vector denoting titles to use on multipanel plots. Must be the same length as processing_method. Default (NULL) uses processing_method character vector for plot titles.}

\item{return_output}{Logical. If true, returns output in a data.frame.}
}
\description{
Make multipanel plots to compare methods/workflows for processing CTD data.
}
