% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_netcdf.R
\name{make_oce_table}
\alias{make_oce_table}
\title{Write cast data to an oce or rds file}
\usage{
make_oce_table(
  cast_files,
  precision = c(temperature = 4, conductivity = 6, salinity = 4, absolute_salinity = 4,
    sound_speed = 3, density = 3, pressure = 3, oxygen = 4, ph = 3, velocity = 3),
  output_file
)
}
\arguments{
\item{cast_files}{Paths to accepted profiles.}

\item{precision}{Precision to use for variables as a named numeric vector.}

\item{output_file}{Output filepath for netCDF file.}
}
\description{
Write cast data to an oce or rds file
}
