% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_flag_interpolate.R
\name{manual_flag_interpolate}
\alias{manual_flag_interpolate}
\title{Manually flag and interpolate bad data}
\usage{
manual_flag_interpolate(csv_paths = NULL, review = c("density"))
}
\arguments{
\item{csv_paths}{Path to csv files to be flagged and reviewed.}
}
\description{
Visually inspect plots of depth versus salinity, and depth versus pressure. Flag data for removal in a GUI and interpolate missing data. Write interpolated data to output/manual_flag
}
