% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_flag_review.R
\name{manual_flag_review}
\alias{manual_flag_review}
\title{Review manual flags}
\usage{
manual_flag_review(var = c("salinity", "temperature"), flag_dir = NULL)
}
\arguments{
\item{var}{Variable to search for point removal.}

\item{flag_dir}{Manual flag directory file path. Default NULL uses the directory [working_directory]/output/manual_flag/}
}
\description{
Review plots from gapctd::manual_flag_interpolate.
}
