% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{median_filter}
\alias{median_filter}
\title{Median window filter (R workflow)}
\usage{
median_filter(
  x,
  variables = c("temperature", "conductivity"),
  window = c(5, 5)
)
}
\arguments{
\item{x}{oce object}

\item{variables}{Character vector of data variable names to filter.}

\item{window}{Numeric vector of windows (i.e. number of scans to use to calculate medians)}
}
\value{
oce object with channels filtered
}
\description{
Applies a median window filter to variables.
}
