% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_bad_rds.R
\name{move_bad_rds}
\alias{move_bad_rds}
\title{Remove rds files based on QA/QC flags (R workflow)}
\usage{
move_bad_rds(rds_dir_path = here::here("output", "gapctd"))
}
\arguments{
\item{rds_dir_path}{File path to directory containing rds files to be evaluated.}
}
\value{
Bad files moved to /bad_cnv/
}
\description{
Remove deployment rds files that contain bottom, upcast, and/or bottom oce objects based on QA/QC flags.
}
