% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{move_binned_cnv}
\alias{move_binned_cnv}
\title{Function to move final .cnv files to bad_cnv and final_cnv directories (SBEDP workflow)}
\usage{
move_binned_cnv()
}
\description{
Moves binned .cnv files without data from the cnv folder to bad_cnv, and files with data to final_cnv.
}
