% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_to_final_cnv.R
\name{move_to_final_cnv}
\alias{move_to_final_cnv}
\title{Move best method files to final_cnv}
\usage{
move_to_final_cnv(best_df)
}
\arguments{
\item{best_df}{Output of gapctd::compare_methods()}
}
\description{
Move best method files based on spikiness of profiles.
}
