% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm_functions.R
\name{optim_ctm_pars}
\alias{optim_ctm_pars}
\title{Optimize cell thermal mass correction parameters (R workflow)}
\usage{
optim_ctm_pars(
  dc = NULL,
  uc = NULL,
  optim_method = "L-BFGS-B",
  start_alpha_C = c(0.001, 0.01, 0.02, 0.04, 0.08, 0.12),
  start_beta_C = c(1, 1/2, 1/4, 1/8, 1/12, 1/24),
  default_parameters = c(alpha_C = 0.04, beta_C = 0.125),
  area_method = "ts"
)
}
\arguments{
\item{dc}{downcast oce object}

\item{uc}{upcast oce object}

\item{optim_method}{Optimization method for optim(). Default is the Broyden-Fletcher-Goldfarb-Shanno with constraints algorithm ("L-BFGS-B")}

\item{start_alpha_C}{Starting value for alpha in cell thermal mass optimization (default = 0.04, typical value for SBE19plus).}

\item{start_beta_C}{Starting value for beta in cell thermal mass optimization (default = 1/8, typical value for SBE19plus).}

\item{default_parameters}{Named numeric vector of default parameters. Defaults is c(alpha_C = 0.04, beta_C = 0.125).}

\item{area_method}{Area between temperature-salinity ("ts") curves, depth-salinity curves ("zs"), or pressure-salinity curves ("ps")}
}
\value{
A named numerical vector of the optimal alpha_C or beta_C. The input values are returned if the optimization does not converge.
}
\description{
Estimate optimal cell thermal mass correction parameters by minimizing the area between downcast and upcast temperature-salinity curves. Or, if only upcast or downcast is provided, optimization based on minimizing the gradient of the salinity profile
}
\author{
Sean Rohan
}
