% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{oxygen_saturation}
\alias{oxygen_saturation}
\title{Calculate oxygen saturation (ml/l)}
\usage{
oxygen_saturation(temperature, salinity)
}
\arguments{
\item{temperature}{Temperature in degrees Celsius}

\item{salinity}{salinity in PSU (PSS-78)}
}
\description{
Calculate oxygen saturation as a function of temperature and salinity based on oxygen solubility from Garcia and Gordon (1992).
}
\references{
Garcia, H.E., Gordon, L.I., 1992. Oxygen solubility in seawater: Better fitting equations. Limnol. Oceanogr. 37, 1307–1312. https://doi.org/10.4319/lo.1992.37.6.1307
}
