% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfit.R
\name{ph_calibration}
\alias{ph_calibration}
\title{Fit pH calibration parameters and generate new configuration file
 
This function estimates pH calibration parameters using Sea-Bird's calibration equation.}
\usage{
ph_calibration(
  xmlcon_in,
  ph_voltage,
  ph,
  temperature,
  ph_serial_number,
  xmlcon_out = NULL
)
}
\arguments{
\item{ph_voltage}{Measured pH channel voltage at the reference standard}

\item{ph}{ph of the reference standard}

\item{temperature}{Reference standard temperature in degrees celcisus}

\item{xmlcon_in_path}{Filepath to the current CTD configuration file.}

\item{xmlcon_out_pth}{Filepath to the output CTD configuration file.}
}
\description{
Fit pH calibration parameters and generate new configuration file
 
This function estimates pH calibration parameters using Sea-Bird's calibration equation.
}
\examples{

# Example of pH calibration with a demo file

ph_calibration(xmlcon_in = system.file("./extdata/example/SBE19plusV2_8106_ph_DO_leg2.xmlcon", package = "gapctd"),
               ph_serial_number = 9090,
               ph_voltage = c(1.977, 2.797, 3.534),
               ph = c(4, 7, 10),
               temperature = 23)
}
\author{
Sean Rohan
}
