% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfit.R
\name{phfit}
\alias{phfit}
\title{This function fits pH calibration parameters
 
This function estimates pH calibration parameters using Sea-Bird's calibration equation.}
\usage{
phfit(ph, ph_voltage, temperature)
}
\arguments{
\item{ph}{ph of the reference standard}

\item{ph_voltage}{Measured pH channel voltage at the reference standard}

\item{temperature}{Reference standard temperature in degrees celcisus}
}
\description{
This function fits pH calibration parameters
 
This function estimates pH calibration parameters using Sea-Bird's calibration equation.
}
\examples{
# Example calibration using three pH reference standards (4, 7, 10)
# Temperature = 23 degrees C
# Measured voltage (e.g. from Real-Time Data Acquisition in Seasave V7)
# See also: SBE pH Sensor Calibration. https://www.youtube.com/watch?v=4uMc5H66dVQ

phfit(
  ph_voltage = c(1.977, 2.797, 3.534),
  ph = c(4, 7, 10),
 temperature = 23)
}
\references{
Sea-Bird Electronics Application Note No. 18-1. SBE 18, 27, and 30, and AMT pH Sensor Calibration (PHFIT Version 2.1) https://www.seabird.com/asset-get.download.jsa?code=250995
}
