% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_calcs.R
\name{profile_bld}
\alias{profile_bld}
\title{Calculate bottom layer depth using a threshold method}
\usage{
profile_bld(rho, z, totdepth, threshold = 0.1, ref.depth = 5)
}
\arguments{
\item{rho}{Numeric vector of densities}

\item{z}{Numeric vector of depths. Depths are positive.}

\item{totdepth}{Maximum depth sampled by the cast}

\item{threshold}{Density threshold}

\item{ref.depth}{Thickness of bottom layer to use for calculating bottom layer density}
}
\description{
Calculate bottom layer depth using a threshold method
}
