% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_calcs.R
\name{profile_bld_from_t}
\alias{profile_bld_from_t}
\title{Calculate bottom layer depth from temperature using a threshold}
\usage{
profile_bld_from_t(
  temperature,
  z,
  ref_dist_from_max = 4,
  min_temperature_diff = 0.5,
  temp_threshold = 0.25,
  na_rm = TRUE
)
}
\arguments{
\item{temperature}{Numeric vector of temperatures}

\item{z}{Numeric vector of depths for each of the temperatures}

\item{ref_dist_from_max}{Numeric vector of distances from the maximum depth to use for the reference temperature.}

\item{min_temperature_diff}{Minimum temperature difference. If the difference in temperatures between the reference depths and shallow depths. If the difference is below this threshold, the function returns the minimum depth (i.e. column is considered fully mixed).}

\item{temp_threshold}{Threshold temperature difference.}

\item{na_rm}{Should NA values be omitted?}
}
\description{
Based on the temperature recorded at a specified difference from the maximum depth for a profile.
}
\author{
Sean Rohan
}
