% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_calcs.R
\name{profile_cokelet_density_diff}
\alias{profile_cokelet_density_diff}
\title{Calculate density difference between reference depth and 30 m below mixed-layer depth (or bottom)}
\usage{
profile_cokelet_density_diff(rho, z, mld, ref.depth = 5, mld.buffer = 30)
}
\arguments{
\item{rho}{Density vector}

\item{z}{Depth vector (positive depths)}

\item{mld}{Vector of mixed layer depth}

\item{ref.depth}{Reference depth for calculating mixed layer density.}

\item{mld.buffer}{Depth difference between upper and lower}
}
\description{
Calculate density difference between reference depth and 30 m below mixed-layer depth (or bottom)
}
