% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_calcs.R
\name{profile_mld_from_t}
\alias{profile_mld_from_t}
\title{Calculate mixed layer depth from temperature using a threshold}
\usage{
profile_mld_from_t(
  temperature,
  z,
  reference_depth = 5,
  min_temperature_diff = 0.5,
  temp_threshold = 0.25,
  assign_inversion
)
}
\arguments{
\item{temperature}{Numeric vector of temperatures}

\item{z}{Numeric vector of depths for each of the temperatures}

\item{reference_depth}{Numeric vector of reference depth, can either be a single depth or multiple depths.}

\item{min_temperature_diff}{Minimum temperature difference. If the difference in temperatures between the reference depths and other depths is below this threshold. If the difference is below this threshold, the function returns the maximum depth (i.e. column is considered fully mixed).}

\item{temp_threshold}{Threshold temperature difference}

\item{assign_inversion}{Substitutes maximum depth when temperature deep in the water column is higher than reference. Rare occurrence that may be an artifact of the survey sampling scheme.}
}
\description{
Based on the temperature recorded at a specified reference depth, calculates the mixed layer depth using only temperature.
}
