% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_calcs.R
\name{profile_pycnocline}
\alias{profile_pycnocline}
\title{Calculate pycnocline depth as the maximum rate of increase in density at depths below the pycnocline and below the reference depth}
\usage{
profile_pycnocline(rho, z, mld = NULL, ref.depth = 5)
}
\arguments{
\item{rho}{Numeric vector of density}

\item{z}{Numeric vector of depths}
}
\description{
Find the depth of the pycnocline as the depth where density is changing the fastest
}
