% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_calcs.R
\name{profile_tsrho_by_layer}
\alias{profile_tsrho_by_layer}
\title{Calculate temperature, salinity, and density for each layer}
\usage{
profile_tsrho_by_layer(t, s, rho, z, mld, mld.buffer)
}
\arguments{
\item{t}{temperature vector}

\item{s}{Salinity vector}

\item{rho}{Density vector}

\item{z}{Depth vector}

\item{mld}{Mixed layer depth}
}
\description{
Calculate temperature, salinity, and density for each layer
}
