% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_check.R
\name{qc_check}
\alias{qc_check}
\title{Check profile for completeness (R workflow)}
\usage{
qc_check(x, prop_max_flag = 0.1, prop_min_bin = 0.875, time_diff_max = 45)
}
\arguments{
\item{x}{oce object}

\item{prop_max_flag}{Proportion of bins with flagged/interpolated data. Default = 0.1}

\item{prop_min_bin}{Minimum proportion of depth bins with data relative to depth. Default = 0.875}

\item{time_diff_max}{Maximum allowable time difference between on-bottom event and cast time.}
}
\value{
Returns oce object with metadata$flags field updated with flags
}
\description{
Assign quality control/quality assurance flag to profile based on whether the data are complete.
}
