% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ctm.R
\name{remedial_ctm}
\alias{remedial_ctm}
\title{UI and wrapper for split_ctm}
\usage{
remedial_ctm(
  rds_path = here::here("output", "gapctd"),
  haul_df,
  threshold = -1e-05
)
}
\arguments{
\item{rds_path}{Path to directory containing rds files.}

\item{haul_df}{data.frame containing haul data from RACEBASE that includes metadata for the cnv file.}

\item{threshold}{Numerical. Threshold for flagging a density inversion. Must be negative. Default (-1e-5) is the buoyancy frequency threshold PMEL uses.}
}
\value{
Profiles (in "_full.rds" files) processed using CTM parameters estimated for individual casts if cell thermal mass correction parameter estimation converged for the individual casts.
}
\description{
Visually inspect plots and decide if raw files should be reprocessed using cell thermal mass correction parameters for individual profiles instead of the parameters estimated for both casts. Profiles selected for additional CTM are processed using split_ctm.
}
\author{
Sean Rohan
}
