% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{remove_bad_station_data}
\alias{remove_bad_station_data}
\title{Filter out bad bottom and surface data and cnvs (SBEDP workflow)}
\usage{
remove_bad_station_data(
  haul_metadata_path = list.files(paste0(getwd(), "/metadata/"), full.names = TRUE),
  max_diff_bt_ctd_depth = 10,
  min_ctd_depth = 5,
  region,
  vessel,
  year
)
}
\arguments{
\item{haul_metadata_path}{Character vector. Path to haul data .csv generated by get_bottom_data().}

\item{max_diff_bt_ctd_depth}{Numeric vector (1L). Maximum permissable difference between BT and CTD depth. Default = 10 m.}

\item{min_ctd_depth}{Numeric vector (1L). Minimum haul depth in meters. Default = 5 m.}

\item{region}{Required. Region as a character vector. Either "BS", "AI", or "GOA".}

\item{vessel}{Required. Vessel number as a numeric vector.}

\item{year}{Required. Year as a numeric vector.}
}
\description{
Filter out data bottom and surface averages that show evidence of operational error and equipment failure (e.g. CTD on while on deck, not reaching haul depth). Run after calc_surface_mean().
}
