% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_profiles.R
\name{review_oxygen_ph_profiles}
\alias{review_oxygen_ph_profiles}
\title{Review oxygen and pH}
\usage{
review_oxygen_ph_profiles(
  rds_dir_path,
  in_pattern = "_final_ts.rds",
  out_pattern = "_final.rds"
)
}
\arguments{
\item{rds_dir_path}{Filepath to directory containing rds files to be reviewed.}

\item{in_pattern}{Character vector search pattern for input files.}

\item{out_pattern}{Character vector to append to filename. Default = "_final.rds"}
}
\description{
Run after selecting temperature/salinity/density profiles.
}
\author{
Sean Rohan
}
