% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_alignment_calcs.R
\name{run_alignment_calcs}
\alias{run_alignment_calcs}
\title{Wrapper function around estimate_alignment}
\usage{
run_alignment_calcs(
  profile_files = sort(c(list.files(here::here("output", "sbe19plus_v0"), full.names =
    TRUE, pattern = "downcast.cnv"), list.files(here::here("output", "sbe19plus_v0"),
    full.names = TRUE, pattern = "upcast.cnv"))),
  make_diagnostic_plots = TRUE,
  min_pressure = 4,
  ...
)
}
\arguments{
\item{profile_files}{Paths to profile files for which alignments should be estimated, as a character vector. For trawl data, should have separate upcast and downcast files.}

\item{make_diagnostic_plots}{Logical. Should diagnostic plots be saved to plots/alignment.pdf?}

\item{min_pressure}{Minimum pressure to use for alignment correlation. It is advisable to exclude near-surface values due to spikes in conductivity, espeically in downcasts.}

\item{...}{Optional arguments passed to gapctd::estimate_alignment().}
}
\description{
Estimate alignment for a channel
}
