% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm_functions.R
\name{run_ctm_adjust_tsarea}
\alias{run_ctm_adjust_tsarea}
\title{Wrapper function around ctm_adjust_tsarea (SBEDP workflow)}
\usage{
run_ctm_adjust_tsarea(
  profile_files = sort(c(list.files(here::here("output", "sbe19plus_v0"), full.names =
    TRUE, pattern = "downcast.cnv"), list.files(here::here("output", "sbe19plus_v0"),
    full.names = TRUE, pattern = "upcast.cnv"))),
  min_pressure = 4,
  optim_method = "L-BFGS-B",
  optim_maxit = 500,
  start_alpha = 0.04,
  start_tau = 8,
  obj_method = "area",
  ...
)
}
\arguments{
\item{profile_files}{Paths to profile files for which CTM parameters should be estimated.}

\item{min_pressure}{Minimum pressure to use for alignment correlation. It is advisable to exclude near-surface values due to spikes in conductivity, especially in downcasts.}

\item{optim_method}{Optimization method for optim(). Default is the Broyden-Fletcher-Goldfarb-Shanno algorithm ("BFGS")}

\item{optim_maxit}{Number of optim iterations or maximum number of iterations, depending on the optim method. Default = 500.}

\item{start_alpha}{Default alpha value.}

\item{start_tau}{Default tau value.}
}
\value{
Returns a data frame containing \itemize{
\item deploy_id: deployment file ID
\item cast_direction: Direction of the cast ("downcast" or "upcast")
\item default_obj: estimated CTM objective function for the default parameters (alpha = 0.04, tau = 8)
\item adj_obj: estimated CTM objective function for the estimated parameters
\item alpha: estimated alpha parameter
\item tau: estimated tau[CTM] parameter
\item mean_t_rate: average rate of change in temperature for the cast in degrees Celsius per db (pressure).
\item mean_s_rate: average rate of change in salinity for the cast in PSS-78 salinity per db (pressure).
\item max_t_rate:  maximum rate of change in temperature for the cast in degrees Celsius per db (pressure).
\item max_s_rate:  maximum rate of change in salinity for the cast in PSS-78 salinity per db (pressure).
\item range_t: absolute range of temperature in a cast in degrees Celsius.
\item range_s: absolute range of salinity in a cast in PSS-78 salinity.}
}
\description{
Estimate conductivity cell thermal inertia correction using ctm_adjust_tsarea and bbmle::mle2().
}
