% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{run_method}
\alias{run_method}
\title{Process files using a specified batch processing method (SBEDP workflow)}
\usage{
run_method(
  vessel,
  year,
  region,
  channel,
  processing_method,
  storage_directory = NULL,
  last_pattern = "TEOS10.cnv",
  ctd_dir = "G:/RACE_CTD/data/2021/ebs/v162",
  alignment_df = NULL,
  ctm_df = NULL,
  ...
)
}
\arguments{
\item{vessel}{RACEBASE Vessel. Passed to gapctd::run_sbe_batch() and gapctd::remove_bad_station_data()}

\item{year}{Year. Passed to gapctd::run_sbe_batch() and gapctd::remove_bad_station_data()}

\item{region}{RACEBASE region. Passed to gapctd::run_sbe_batch() and gapctd::remove_bad_station_data()}

\item{channel}{RODBC channel. Passed to gapctd::run_sbe_batch(), and gapctd::get_haul_events(), gapctd::remove_bad_station_data()}

\item{processing_method}{Character vector indicating which processing method to use. Passed to gapctd::setup_ctd_processing_dir}

\item{ctd_dir}{Directory containing CTD data. Passed to gapctd::setup_ctd_processing_dir}

\item{rm_files}{Logical. Should be removed from the directory before starting processing}
}
\description{
Function to run a batch processing workflow and save output. Useful for comparing batch processing workflows using different sets of SBE data processing modules.
}
