% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sbe_batch.R
\name{run_sbe_batch}
\alias{run_sbe_batch}
\title{Function to run SBE19 batch file}
\usage{
run_sbe_batch(xmlcon_file = NA, bat_file = NA, derive_file = NA)
}
\arguments{
\item{xmlcon_file}{Optional character vector specifying the name of the .xmlcon file. Default NA automatically retrieves the filename from psa_xmlcon subdirectory of the working directory.}

\item{bat_file}{Optional character vector specifying the name of the .bat file in the working directory to use. Must specify if multiple batch files are present in the working directory (i.e., you're comparing multiple processing methods).}
}
\description{
Wrapper function to run SBE DataProcessing on Windows using command line in the local directory. Run after gapctd::setup_ctd_processing_dir().
}
