% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_profiles.R
\name{select_best_method}
\alias{select_best_method}
\title{Select best processing method (R workflow)}
\usage{
select_best_method(rds_dir_path)
}
\arguments{
\item{rds_dir_path}{Filepath to directory containing .rds files.}
}
\description{
Simultaneously review multiple temperature, salinity, and density profiles from downcasts and upcasts that were contained in .rds files that were processed using different methods (e.g., Typical, Typical CTM, TSA, MSG) then select the best profile. Renames the rds file with the best profile by replacing the method with the suffix ("_best.rds"). Run after wrapper_run_gapctd().
}
