% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ctd_processing_dir.R
\name{setup_ctd_processing_dir}
\alias{setup_ctd_processing_dir}
\title{Function to copy CTD files from survey directory to a working directory}
\usage{
setup_ctd_processing_dir(ctd_dir, output_dir = NA, ctd_unit)
}
\arguments{
\item{ctd_dir}{Character vector. indicating the filepath to the CTD directory.}

\item{output_dir}{Character vector indicating directory for saving the output. Default NA uses the current directory.}

\item{unit}{Character vector indicating model of the unit. Options = sbe19plus, sbe19. Won't run if the wrong unit is used.}
}
\description{
Function to copy CTD files from survey directory to a working directory
}
