% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{slowdown}
\alias{slowdown}
\title{Flag slowdowns and reversals (R workflow)}
\usage{
slowdown(
  x,
  min_speed = 0.1,
  window = 5,
  cast_direction = NULL,
  exclude_bottom = 2
)
}
\arguments{
\item{x}{oce object}

\item{min_speed}{Numeric vector (1L). Threshold below which observations should be flagged.}

\item{window}{Numeric vector (1L). Number of samples to use to calculate speeds for each observation.}

\item{cast_direction}{Character vector ("downcast", "upcast") denoting the cast direction.}

\item{exclude_bottom}{Pressure width to exclude near-bottom. This is used because it can takes awhile for the CTD to start ascending when the net approaches or leaves the seafloor, but the contact between the net and the seafloor generally does not cause influential flow reversals.}
}
\value{
oce object with flags updated to denote scans for which speeds were below the threshold.
}
\description{
Flags slowdowns and reversals in profiles based on minimum speed.
}
\author{
Sean Rohan
}
