% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ctm.R
\name{split_ctm}
\alias{split_ctm}
\title{Remedial method to fix CTM errors using the best upcast or downcast parameter}
\usage{
split_ctm(rds_path, haul_df)
}
\arguments{
\item{rds_path}{Path to rds file}

\item{haul_df}{data.frame containing haul data from RACEBASE that includes metadata for the cnv file.}
}
\value{
An upcast and downcast rds file if both casts were found in the rds file. Moves the original rds file to bad_cnv.
}
\description{
Reads-in an rds file containing data that has been processed using the full gapctd workflow and uses the file name and estimated CTM parameters to reprocess data using the best upcast and downcast parameters.
}
