% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{tau_par}
\alias{tau_par}
\title{Tau correction for dissolved oxygen voltage}
\usage{
tau_par(temperature, pressure, tau20, d0, d1, d2)
}
\arguments{
\item{temperature}{Temperature in degrees C}

\item{pressure}{Pressure in dbar}

\item{tau20}{Tau correction calibration parameter Tau20.}

\item{d0}{Tau correction calibration parameter D0.}

\item{d1}{Tau correction calibration parameter D1.}

\item{d2}{Tau correction calibration parameter D2.}
}
\description{
Tau correction following Edwards et al. (2010).
}
\references{
Edwards, B., Murphy, D., Janzen, C., Larson, A.N., 2010. Calibration, response, and hysteresis in deep-sea dissolved oxygen measurements. J. Atmos. Ocean. Technol. 27, 920–931. https://doi.org/10.1175/2009JTECHO693.1
}
