% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{temperature_correction}
\alias{temperature_correction}
\title{Temperature sensor thermal inertia correction (R workflow)}
\usage{
temperature_correction(x, alpha_T, beta_T, freq_n)
}
\arguments{
\item{x}{oce object that includes timeS, temperature, and conductivity channels.}

\item{alpha_T}{Alpha parameter for temperature correction (1L numeric)}

\item{freq_n}{Optional. Sampling interval in seconds (1L numeric). Sampling interval is inferred from timeS if not provided.}

\item{beta_C}{Beta parameter for conductivity correction (1L numeric)}
}
\value{
oce object with correction applied to temperature.
}
\description{
Discrete time filter to correct thermal inertia errors in temperature sensor measurements.
}
