% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm_functions.R
\name{ts_area}
\alias{ts_area}
\title{Calculate area between temperature-salinity curves (R workflow)}
\usage{
ts_area(dc, uc, by = "pressure", return_sf = FALSE, area_method = "ts")
}
\arguments{
\item{dc}{downcast oce object}

\item{uc}{upcast oce object}

\item{by}{Which Z dimension variable should be used ("depth" or "pressure")?}

\item{return_sf}{Logical. If TRUE, returns sf object with polygons. Otherwise returns the sum of polygon areas as a 1L numeric vector.}

\item{area_method}{Area between temperature-salinity ("ts") curves or oxygen-pressure ("op") curves.}
}
\value{
When return_sf = TRUE, an sf object (POLYGON) of the area between T-S curves. When return_sf = FALSE, a 1L numeric vector with the sum of polygon areas.
}
\description{
Calculate area between temperature-salinity curves (R workflow)
}
