% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_check.R
\name{wrapper_flag_interpolate}
\alias{wrapper_flag_interpolate}
\title{Wrapper for qc_flag_interpolate (R workflow)}
\usage{
wrapper_flag_interpolate(
  rds_dir_path = here::here("output", "gapctd"),
  output_dir_path = NULL,
  append_char = "_qc",
  review = "density",
  bin_var = "depth"
)
}
\arguments{
\item{rds_dir_path}{Filepath to directory containing rds files to be reviewed.}

\item{append_char}{Characters to append to output file name.}

\item{review}{Passed to qc_flag_interpolate(). Variable to review ("density", "temperature" or "salinity"). Default = "density"}

\item{output_path}{Optional. Filepath to output directory. If not provided, files are written to the rds_dir_path directory.}
}
\value{
rds files with flagged conductivity and temperature removed and interpolated; derived quantities recalculated.
}
\description{
Visually inspect and interpolate errors. See documentation for gapctd:::qc_flag_interpolate()
}
