% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gapctd.R
\name{wrapper_run_gapctd}
\alias{wrapper_run_gapctd}
\title{Wrapper function for run_ctd (R workflow)}
\usage{
wrapper_run_gapctd(
  cnv_dir_path = here::here("cnv"),
  processing_method,
  return_stage = "full",
  gapctd_round = 1,
  haul_df = NULL,
  vessel = NULL,
  cruise = NULL,
  channel = NULL,
  ctm_pars = NULL,
  align_pars = NULL
)
}
\arguments{
\item{cnv_dir_path}{Path to directory containing cnv files to be processed.}

\item{processing_method}{Name of processing method; for saving output.}

\item{return_stage}{Character vector denoting which stages of processing should be included in the output (options "typical", "split", "align", "tmcorrect", "full")}

\item{gapctd_round}{Processing round (1 = typical settings, 2 = optimize CTM w/ T-S area or S path distance.}

\item{haul_df}{Optional. data.frame containing  haul data from RACEBASE. Must provide arguments for vessel and cruise if not provided.}

\item{vessel}{Optional. Vessel code as a 1L numeric vector.}

\item{cruise}{Optional. Cruise code as a numeric vector (>= 1L).}

\item{channel}{Optional. RODBC channel; only used when haul_df = NULL.}

\item{ctm_pars}{Optional. Used for remedial cell thermal mass corrections. Optional list of parameters to use for cell thermal mass correction. Must contain alpha_C and beta_C.}

\item{align_pars}{Optional. A list object with alignment parameters for a variable, e.g., list(temperature = -0.5)}
}
\value{
Writes rds files with cast data to /output/[processing_method]
}
\description{
Runs run_ctd() on all cnv files in a filepath.
}
