% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{convert_ctd_hex}
\alias{convert_ctd_hex}
\title{Check at-sea CTD data}
\usage{
convert_ctd_hex(
  ctd_directory_path,
  xmlcon_path,
  bat_file = NULL,
  datcnv_file = NULL
)
}
\description{
In development
 
 @param ctd_directory_path Path to CTD directory
 @param xmlcon_path Path to xmlcon file
 @noRd
}
