% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{convert_ctd_hex}
\alias{convert_ctd_hex}
\title{Convert hex files using SBE Data Processing software}
\usage{
convert_ctd_hex(
  hex_file_path,
  xmlcon_path,
  bat_file = NULL,
  datcnv_file = NULL
)
}
\arguments{
\item{xmlcon_path}{Path to xmlcon file}

\item{ctd_directory_path}{Path to CTD directory}
}
\description{
Converts .hex files to .cnv files using SBE Data Processing. Must have SBE Data Processing software installed for this function to work.
}
