% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NMEA_files.R
\name{create_NMEA_files}
\alias{create_NMEA_files}
\title{Function to pull in lat/lon from bottom trawl racebase haul data that matches each .cnv to create NMEA files}
\usage{
create_NMEA_files()
}
\arguments{
\item{oracle_username}{Character. Enter your personal oracle username to access RACEBASE.HAUL oracle data}

\item{oracle_password}{Character. Enter your personal oracle password to access RACEBASE.HAUL oracle data}

\item{survey_region}{Character. Specify the survey region of your CTD data (GOA, AI, BS). Prompted, no default.}

\item{survey_year}{Character. Specify the survey year of your CTD data. Prompted, no default.}
}
\description{
implement between derive.bat and getdata.bat. Requires a VPN connection to run. 
Function will prompt for oracle username, oracle password, survey region (GOA, AI, BS), and survey year.
Function dependencies: getPass, RODBC, oce
}
