% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NMEA_files.R
\name{create_NMEA_files}
\alias{create_NMEA_files}
\title{Create NMEA files with LAT/LON for Derive TEOS-10}
\usage{
create_NMEA_files(
  rodbc_channel = NA,
  vessel,
  region,
  year,
  haul_df = NULL,
  write_metadata = TRUE
)
}
\arguments{
\item{rodbc_channel}{Required if haul_csv is not provided. Must provide an open RODBC channel (this parameter) or path to haul_csv. Default = NA.}

\item{vessel}{Required. Vessel number as a numeric vector.}

\item{region}{Required. Region as a character vector. Either "bs", "ai", or "goa".}

\item{year}{Required. Year as a numeric vector.}

\item{haul_df}{If TRUE, loads an rds file named "data/haul_data.rds" instead of running a query to retrieve haul data.}
}
\description{
Function to retrieve haul data from RACEBASE or a csv file and generate NMEA (.txt) files that match .cnv file names. Function dependencies: getPass, RODBC, oce
}
