% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{get_calibration_parameter}
\alias{get_calibration_parameter}
\title{Extract variable from xmlcon file text}
\usage{
get_calibration_parameter(
  header,
  cal_par,
  start_block = NULL,
  end_block = NULL,
  make_numeric = TRUE
)
}
\arguments{
\item{header}{Character vector of lines from an xmlcon file}

\item{cal_par}{Character vector of the xmlcon tag}

\item{start_block}{starting index for lines of the header to search}

\item{end_block}{ending index for lines of the leader to search}

\item{make_numeric}{Logical. Should the tag value be forced to a numeric?}
}
\description{
Extract variable from xmlcon file text
}
