% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{hex_to_cnv}
\alias{hex_to_cnv}
\title{Convert SBE19plus V2 hex files to cnv}
\usage{
hex_to_cnv(
  hex_path,
  output_path,
  xmlcon_path = NULL,
  sample_interval = 0.25,
  output_channels = NULL,
  output_sig_digits = NULL
)
}
\arguments{
\item{hex_path}{Path to a .hex file}

\item{output_path}{Path to the output file location for a .cnv file}

\item{xmlcon_path}{Optional. Path to config file. Must be provided if .hex file does not contain configuration file parameters.}

\item{sample_interval}{Sampling interval for scans; 0.25 for a typical SBE19plus V2 deployment.}

\item{output_channels}{Optional. Named vector of output channels and their names. Do not use unless outputs are are not the defaults.}

\item{output_sig_digits}{Optional. Significant digits after the decimal place for output channels. Only change if a subset of channels. Do not use unless outputs are are not the defaults.}
}
\description{
This function decodes hexadecimal-formatted Sea-Bird CTD files to cnv files.
}
\author{
Sean Rohan
}
