% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{hex_to_cnv}
\alias{hex_to_cnv}
\title{Convert SBE19plus V2 hex files to cnv without SBE Data Processing}
\usage{
hex_to_cnv(
  hex_path,
  output_path,
  xmlcon_path = NULL,
  sample_interval = 0.25,
  output_channels = c(time_elapsed = "timeS: Time, Elapsed [seconds]", temperature =
    "tv290C: Temperature [ITS-90, deg C]", pressure =
    "prdM: Pressure, Strain Gauge [db]", conductivity = "c0S/m: Conductivity [S/m]", flag
    = "flag:  0.000e+00"),
  output_sig_digits = c(time_elapsed = 3, temperature = 4, pressure = 3, conductivity =
    6, flag = 1)
)
}
\arguments{
\item{hex_path}{Path to a .hex file}

\item{output_path}{Path to the output file location for a .cnv file}

\item{sample_interval}{Sampling interval for scans; 0.25 for a typical SBE19plus V2 deployment.}

\item{output_channels}{Named vector of output channels and their names. Do not change from default unless additional channels of data are added.}

\item{output_sig_digits}{Significant digits after the decimal place for output channels. Do not change from default unless additional channels of data are added or sensor precision changes.}

\item{xmlcon_file}{Optional. Path to config file. Must be provided if .hex file does not contain configuration file parameters.}
}
\description{
This function decodes hexadecimal-formatted files that are generated by SeaTerm software then writes converted data to cnv files. Output files are similar to the output of SBE Data Processing software and can be read into R using oce::read.ctd().
}
