% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{integer_to_conductivity}
\alias{integer_to_conductivity}
\title{Convert SBE integer to conductivity}
\usage{
integer_to_conductivity(
  conductivity_integer,
  temperature,
  pressure,
  condg,
  condh,
  condi,
  condj,
  cpcor,
  ctcor,
  par0 = 256,
  par1 = 1000,
  sig_figs = 6
)
}
\arguments{
\item{conductivity_integer}{Conductivity voltage integer}

\item{temperature}{Temperature in degrees C}

\item{pressure}{Presssure in degrees C}

\item{condg}{Conductivity calibration parameter condg}

\item{condh}{Conductivity calibration parameter condh}

\item{condi}{Conductivity calibration parameter condi}

\item{condj}{Conductivity calibration parameter condj}

\item{cpcor}{Conductivity calibration parameter cpcor}

\item{ctcor}{Conductivity calibration parameter ctcor}

\item{par0}{Constant to convert integer to voltage}

\item{par1}{Constant to convert integer to voltage}

\item{sig_figs}{number of significant digits to use for conductivity (default = 6)}
}
\description{
Convert SBE integer to conductivity
}
