% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{integer_to_dissolved_oxygen}
\alias{integer_to_dissolved_oxygen}
\title{Convert SBE integer to dissolved oxygen (ml/l)}
\usage{
integer_to_dissolved_oxygen(
  do_integer,
  doa,
  dob,
  doc,
  doe,
  dosoc,
  temperature,
  par0 = 13107,
  pressure,
  salinity,
  Voffset
)
}
\description{
Convert SBE integer to dissolved oxygen (ml/l)
}
