% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctd_hex.R
\name{integer_to_ph}
\alias{integer_to_ph}
\title{Convert SBE integer to pH}
\usage{
integer_to_ph(
  ph_integer,
  ph_offset,
  ph_slope,
  temperature,
  par0 = 13107,
  sig_figs = 3
)
}
\description{
Convert SBE integer to pH
}
