% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{integer_to_pressure}
\alias{integer_to_pressure}
\title{Convert SBE integer to pressure}
\usage{
integer_to_pressure(
  pressure_integer,
  tvoltage_integer,
  ptempa0,
  ptempa1,
  ptempa2,
  ptca0,
  ptca1,
  ptca2,
  ptcb0,
  ptcb1,
  ptcb2,
  pa0,
  pa1,
  pa2,
  par0 = 13107,
  sig_figs = 3,
  convert_to_dbar = TRUE
)
}
\arguments{
\item{pressure_integer}{Pressure voltage integer}

\item{tvoltage_integer}{Temperature voltage integer}

\item{ptempa0}{Pressure calibration parameter ptempa0}

\item{ptempa1}{Pressure calibration parameter ptempa1}

\item{ptempa2}{Pressure calibration parameter ptempa2}

\item{ptca0}{Pressure calibration parameter ptca0}

\item{ptca1}{Pressure calibration parameter ptca1}

\item{ptca2}{Pressure calibration parameter ptca2}

\item{ptcb0}{Pressure calibration parameter ptcb0}

\item{ptcb1}{Pressure calibration parameter ptcb1}

\item{ptcb2}{Pressure calibration parameter ptcb2}

\item{pa0}{Pressure calibration parameter pa0}

\item{pa1}{Pressure calibration parameter pa1}

\item{pa2}{Pressure calibration parameter pa2}

\item{par0}{tvoltage_integer conversion constant}

\item{sig_figs}{number of significant digits to use for temperature (default = 3)}

\item{convert_to_dbar}{Should pressure be returned in or decibars (TRUE) or pounds per square inch without offset (FALSE)}
}
\description{
Convert SBE integer to pressure
}
\author{
Sean Rohan
}
