% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{integer_to_temperature}
\alias{integer_to_temperature}
\title{Convert SBE integer to ITS-90 temperature}
\usage{
integer_to_temperature(
  temperature_integer,
  sig_figs = NULL,
  a0,
  a1,
  a2,
  a3,
  offset = 0,
  par0 = 524288,
  par1 = 1.6e+07,
  par2 = 2.9e+09,
  par3 = 102400000,
  par4 = 20480,
  par5 = 2e+05
)
}
\arguments{
\item{temperature_integer}{Temperature voltage integer}

\item{sig_figs}{number of significant digits to use for temperature}

\item{a0}{Temperature calibration parameter a0}

\item{a1}{Temperature Calibration parameter a1}

\item{a2}{Temperature calibration parameter a2}

\item{a3}{Temperature calibration parameter a3}

\item{offset}{Temperature calibration parameter offset}

\item{par0}{Temperature calibration parameter par0}

\item{par1}{Temperature calibration parameter par1}

\item{par2}{Temperature calibration parameter par2}

\item{par3}{Temperature calibration parameter par3}

\item{par4}{Temperature calibration parameter par4}

\item{par5}{Temperature calibration parameter par5}
}
\description{
Convert SBE integer to ITS-90 temperature
}
