% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_bad_rds}
\alias{move_bad_rds}
\title{Remove rds files based on QA/QC flags (R workflow)}
\usage{
move_bad_rds(
  rds_dir_path = here::here("output", "gapctd"),
  in_pattern = "_typical.rds"
)
}
\arguments{
\item{rds_dir_path}{File path to directory containing rds files to be evaluated.}

\item{in_pattern}{Character vector search pattern for input files.}
}
\value{
Bad files moved to /bad_cnv/
}
\description{
Remove deployment rds files that contain bottom, upcast, and/or bottom ctd objects based on QA/QC flags.
}
\author{
Sean Rohan
}
