% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gapctd.R
\name{run_gapctd}
\alias{run_gapctd}
\title{Process an oce object using the gapctd workflow (R workflow)}
\usage{
run_gapctd(
  x,
  haul_df,
  return_stage = "full",
  ctd_tz = "America/Anchorage",
  ctm_pars = list(),
  align_pars = c(),
  cal_rds_path = NULL,
  cor_var = "conductivity"
)
}
\arguments{
\item{x}{oce object}

\item{haul_df}{data.frame containing haul data from RACEBASE that includes metadata for the cnv file.}

\item{return_stage}{Character vector denoting which stages of processing should be included in the output (options "typical", "split", "align", "tmcorrect", "full"). Can return multiple stages simultaneously. Default = "full"}

\item{ctd_tz}{timezone for CTD as a character vector or numeric that is valid for POSIXct.}

\item{ctm_pars}{Used for remedial cell thermal mass corrections. Optional list of parameters to use for cell thermal mass correction. Must contain alpha_C and beta_C.}

\item{align_pars}{A list object with alignment parameters for a variable, e.g., list(temperature = -0.5)}

\item{cal_rds_path}{Filepath to RDS containing calibration parameters. Required for oxygen data processing.}

\item{cor_var}{Channel to use for optimizing temperature alignment. For exploratory purposes only.}
}
\value{
A list of oce objects at stages of processing specified in return_stage.
}
\description{
Run gapctd modules in the order described in Rohan et al. (2023) to process data from an oce object.
}
\references{
Rohan, S. K., Charriere, N. E., Riggle, B., O’Leary, C. A., and Raring, N. W. 2023. A flexible approach for processing data collected using trawl-mounted CTDs during Alaska bottom-trawl surveys. U.S. Dep. Commer., NOAA Tech. Memo. NMFS-AFSC-475, 43 p.
}
\author{
Sean Rohan
}
