% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gapctd.R
\name{run_gapctd}
\alias{run_gapctd}
\title{Process a .cnv file using the R workflow (R workflow)}
\usage{
run_gapctd(
  x,
  haul_df,
  return_stages = c("final"),
  ctd_tz = "America/Anchorage"
)
}
\arguments{
\item{x}{oce file}

\item{haul_df}{data.frame containing haul data from RACEBASE that includes metadata for the cnv file.}

\item{return_stages}{Character vector denoting which stages of processing should be included in the output (options "typical", "split", "align", "tmcorrect", "final"). Can return multiple stages simultaneously. Default = "final"}

\item{ctd_tz}{timezone for CTD as a character vector or numeric that is valid for POSIXct.}
}
\value{
A list of oce objects at stages of processing specified in return_stages.
}
\description{
Run all gapctd modules in order to process data from a cnv file.
}
