% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe_modules.R
\name{section_oce}
\alias{section_oce}
\title{Section (R workflow)}
\usage{
section_oce(x, by = "timeS", start = NULL, end = NULL, cast_direction = NULL)
}
\arguments{
\item{x}{oce object}

\item{by}{Sectioning variable as a character vector ("timeS", "scan", "datetime")}

\item{start}{Numeric or POSIXct. Start time or scan to include.}

\item{end}{Numeric or POSIXct. End time or scan to include.}

\item{cast_direction}{Cast direction ("downcast", "upcast", "bottom)}
}
\value{
An oce object that only includes variables that were collected within the time/scan interval.
}
\description{
Extract scans within a time or scan interval from an oce object. Time- is time elapsed since measurements started.
}
