% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_to_nc_class}
\alias{vec_to_nc_class}
\title{Detect R vector class and return equivalent netCDF class}
\usage{
vec_to_nc_class(vec, pkg = "RNetCDF")
}
\arguments{
\item{vec}{A vector of any type.}
}
\description{
Supports numeric, integer, and character classes.
}
