% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_check.R
\name{wrapper_flag_interpolate}
\alias{wrapper_flag_interpolate}
\title{Wrapper for qc_flag_interpolate (R workflow)}
\usage{
wrapper_flag_interpolate(
  rds_dir_path,
  in_pattern = NULL,
  append_pattern = NULL,
  review = c("density", "salinity"),
  bin_var = "depth"
)
}
\arguments{
\item{rds_dir_path}{Filepath to directory containing rds files to be reviewed.}

\item{in_pattern}{Character vector search pattern for input files.}

\item{append_pattern}{Characters to append to output file name.}

\item{review}{Passed to qc_flag_interpolate(). Variable to review ("density", "temperature" or "salinity"). Default = "density"}

\item{bin_var}{Name of variable that was used to bin data ("depth" or "pressure").}
}
\value{
rds files with flagged conductivity and temperature removed and interpolated; derived quantities recalculated.
}
\description{
Visually inspect and interpolate errors. See documentation for gapctd:::qc_flag_interpolate()
}
