% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gapctd.R
\name{wrapper_run_gapctd}
\alias{wrapper_run_gapctd}
\title{Wrapper function for run_ctd (R workflow)}
\usage{
wrapper_run_gapctd(
  cnv_dir_path = here::here("cnv"),
  processing_method,
  haul_df = NULL,
  vessel = NULL,
  cruise = NULL,
  channel = NULL,
  racebase_tzone = "America/Anchorage",
  ctd_tzone = "America/Anchorage"
)
}
\arguments{
\item{cnv_dir_path}{Path to directory containing cnv files to be processed.}

\item{processing_method}{Name of processing method; for saving output.}

\item{haul_df}{Optional. data.frame containing  haul data from RACEBASE. Must provide arguments for vessel and cruise if not provided.}

\item{vessel}{Optional. Vessel code as a 1L numeric vector.}

\item{cruise}{Optional. Cruise code as a numeric vector (>= 1L).}

\item{channel}{Optional. RODBC channel; only used when haul_df = NULL.}

\item{racebase_tzone}{Time zone for events and start_time in racebase/race_data tables. Passed to get_haul_data()}

\item{ctd_tzone}{Time zone for CTD events. Passed to run_gapctd().}
}
\value{
Writes rds files with cast data to /output/[processing_method]
}
\description{
Runs run_ctd() on all cnv files in a filepath.
}
